seperates the file for CiscoStyleCli class
